requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cloudinary==1.30.0
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
oauthlib==3.2.1
Pillow==9.2.0
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23.1
